requirements.txt
Django==1.11.3
django-model-utils==3.0.0
pytz==2017.2
jsonfield==2.0.2
django-pipeline==1.6.13
Fabric==1.13.2
elasticsearch==5.4.0
gunicorn==19.7.1
django-jinja==2.3.1
django-extensions==1.8.1
django-user-agents==0.3.2
django-debug-toolbar==1.8
django-celery==3.2.1
django-redis==4.8.0
django-ckeditor==5.3.0
django-waffle==0.12.0
asn1crypto==0.22.0
cryptography==2.0.3
pycrypto==2.6.1
requests==2.18.4
urllib3>=1.22
requests-oauthlib==0.8.0
Pillow==4.2.1
boto3==1.4.7
django-storages==1.6.3
beautifulsoup4==4.6.0
django-juno-testrunner==0.4.1
mysqlclient==1.3.12
django-ses==0.8.4
django-versatileimagefield==1.8.1
BeautifulSoup==3.2.1
sorl-thumbnail==12.4.1
googlemaps==2.5.1
google-api-python-client==1.6.7
djangorestframework==3.8.2
certifi==2015.04.28
drf-extensions==0.3.1
Collectfast==0.6.2
channels==1.1.5
asgi_redis==1.4.3
websocket-client==0.48.0
django-webpush==0.2.5
arrow==0.12.1
sendgrid==5.4.1